requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
fastapi
uvicorn
python-multipart